fix: mobile header
